requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
discord==2.2.2
discord.py==2.2.2
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
imageio-ffmpeg==0.4.8
multidict==6.0.4
numpy==1.24.2
Pillow==9.5.0
pycord==0.1.1
python-dotenv==1.0.0
yarl==1.8.2
